+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [#074] Fixed 'illegal offset' warning in `show_blog_category()` function. XMLoject is now cast properly as a string to use as array ofset. This prevented `show_blog_categories()` from being able to count the posts.
 - [#076] Fixed language string concatenated use for 'Manage Custom Fields'.
 - [#077] Fixed incorrect language string references for the RSS Auto-Importer Settings page.
+- [#079] Fixed some issues with the CustomFields Manager. Also added a sidebar button so that CFM could actually be used.
 
 **Front End UI Improvements**
 - [#074] Properly implemented post counts in categories. This was incomplete in the previous release. Rebuilt methods to make it more robust than originally planned.